+++ resources/[dg]/dg-admin/typescript/src/server/modules/penaltyPoints/controller.penaltyPoints.ts
@@ -1,5 +1,6 @@
-import { scheduleFallofs } from './service.penaltyPoints';
+import { loadPoints, scheduleFallofs } from './service.penaltyPoints';
 
 setImmediate(() => {
   scheduleFallofs();
+  loadPoints();
 });