+++ resources/[dg]/dg-admin/typescript/src/server/modules/penaltyPoints/service.penaltyPoints.ts
@@ -2,6 +2,7 @@ import { SQL } from '@dgx/server';
 import dayjs from 'dayjs'
 
 const resetTimeouts: Record<string, NodeJS.Timeout> = {};
+const adminPoints: Map<string, number> = new Map();
 
 export const scheduleFallofs = async () => {
   const playerPoints = await SQL.query<Penalty.PenaltyReset[]>(
@@ -22,6 +23,7 @@ export const scheduleFallofs = async () => {
 };
 
 export const updatePointsReset = (steamid: string, addPoints: number) => {
+  if (addPoints < 0) return;
   if (resetTimeouts[steamid]) {
     // no way this timeout is happening in the same server restart
     clearTimeout(resetTimeouts[steamid]);
@@ -30,4 +32,15 @@ export const updatePointsReset = (steamid: string, addPoints: number) => {
     INSERT INTO admin_points (steamid, points) VALUES (?, ?)
     ON DUPLICATE KEY UPDATE points = VALUES(points), updated_at = CURRENT_TIMESTAMP()
   `, [steamid, addPoints]);
+    adminPoints.set(steamid, getPointsForSteamId(steamid) + addPoints)
+}
+
+export const loadPoints = async () => {
+  const pointData = await SQL.query<{points: number, steamid: string}[]>("SELECT points, steamid FROM admin_points");
+  if (!pointData) return;
+  pointData.forEach(pd=> adminPoints.set(pd.steamid, pd.points));
+}
+
+export const getPointsForSteamId = (steamId: string) => {
+  return adminPoints.get(steamId) ?? 0;
 }